# Heavy vector octet optimizations