--- a/adnl/rldp/raptorq/discmath/oct.go
+++ b/adnl/rldp/raptorq/discmath/oct.go
@@ -1,5 +1,7 @@
 package discmath
 
+import "unsafe"
+
 var _LogPreCalc = [...]uint8{0, 0, 1, 25, 2, 50, 26, 198, 3, 223, 51, 238, 27, 104, 199, 75, 4, 100, 224, 14, 52, 141, 239,
 	129, 28, 193, 105, 248, 200, 8, 76, 113, 5, 138, 101, 47, 225, 36, 15, 33, 53, 147, 142, 218, 240,
 	18, 130, 69, 29, 181, 194, 125, 106, 39, 249, 185, 201, 154, 9, 120, 77, 228, 114, 166, 6, 191, 139,
@@ -45,15 +47,11 @@ func calcOctMulTable() [256][256]uint8 {
 
 	for i := 1; i < 256; i++ {
 		for j := 1; j < 256; j++ {
-			result[i][j] = OctMul(uint8(i), uint8(j))
+			result[i][j] = _ExpPreCalc[uint32(_LogPreCalc[i])+uint32(_LogPreCalc[j])]
 		}
 	}
 
 	return result
-}
-
-func OctAddMul(x, y uint8) uint8 {
-	return _MulPreCalc[x][y]
 }
 
 func OctExp(x uint32) uint8 {
@@ -64,10 +62,46 @@ func OctInverse(x uint8) uint8 {
 	return OctExp(uint32(255 - _LogPreCalc[x]))
 }
 
-func OctMul(x, y uint8) uint8 {
-	if x == 0 || y == 0 {
-		return 0
+func OctVecAdd(x, y []byte) {
+	xUint64 := *(*[]uint64)(unsafe.Pointer(&x))
+	yUint64 := *(*[]uint64)(unsafe.Pointer(&y))
+
+	for i := 0; i < len(x)/8; i++ {
+		xUint64[i] ^= yUint64[i]
 	}
 
-	return _ExpPreCalc[uint32(_LogPreCalc[x])+uint32(_LogPreCalc[y])]
+	for i := len(x) - len(x)%8; i < len(x); i++ {
+		x[i] ^= y[i]
+	}
+}
+
+func OctVecMul(vector []byte, multiplier uint8) {
+	table := _MulPreCalc[multiplier]
+	for i := 0; i < len(vector); i++ {
+		vector[i] = table[vector[i]]
+	}
+}
+
+func OctVecMulAdd(x, y []byte, multiplier uint8) {
+	table := _MulPreCalc[multiplier]
+	xUint64 := *(*[]uint64)(unsafe.Pointer(&x))
+	pos := 0
+	for i := 0; i < len(x)/8; i++ {
+		var prod uint64
+		prod |= uint64(table[y[pos]])
+		prod |= uint64(table[y[pos+1]]) << 8
+		prod |= uint64(table[y[pos+2]]) << 16
+		prod |= uint64(table[y[pos+3]]) << 24
+		prod |= uint64(table[y[pos+4]]) << 32
+		prod |= uint64(table[y[pos+5]]) << 40
+		prod |= uint64(table[y[pos+6]]) << 48
+		prod |= uint64(table[y[pos+7]]) << 56
+
+		pos += 8
+		xUint64[i] ^= prod
+	}
+
+	for i := len(x) - len(x)%8; i < len(x); i++ {
+		x[i] ^= table[y[i]]
+	}
 }