Old version:
```
Benchmark_EncodeDecodeFuzz-10         1790     611783 ns/op   652920 B/op     1810 allocs/op
Benchmark_EncodeDecodeFuzz-10         1803     600854 ns/op   652919 B/op     1810 allocs/op
Benchmark_EncodeDecodeFuzz-10         1818     619967 ns/op   652914 B/op     1809 allocs/op
```

New version:
```
Benchmark_EncodeDecodeFuzz-10         4076     252432 ns/op   652919 B/op     1810 allocs/op
Benchmark_EncodeDecodeFuzz-10         4021     272682 ns/op   652918 B/op     1810 allocs/op
Benchmark_EncodeDecodeFuzz-10         4015     252178 ns/op   652919 B/op     1810 allocs/op
```

**Result: ~ +125% perfomance increase**